fix: Avoid WriteKind duplication between connector and logical plan
Acutally I think it will be better to logical plan depends on connector, but it is as it is
