+++ axiom/logical_plan/LogicalPlanNode.cpp
@@ -395,21 +395,4 @@ void TableWriteNode::accept(
   visitor.visit(*this, context);
 }
 
-namespace {
-
-folly::F14FastMap<WriteKind, std::string_view> writeKindNames() {
-  static const folly::F14FastMap<WriteKind, std::string_view> kNames = {
-      {WriteKind::kCreate, "CREATE"},
-      {WriteKind::kInsert, "INSERT"},
-      {WriteKind::kUpdate, "UPDATE"},
-      {WriteKind::kDelete, "DELETE"},
-  };
-
-  return kNames;
-}
-
-} // namespace
-
-VELOX_DEFINE_ENUM_NAME(WriteKind, writeKindNames);
-
 } // namespace facebook::axiom::logical_plan
