+++ axiom/optimizer/ToGraph.cpp
@@ -1758,10 +1758,8 @@ void ToGraph::addLimit(const lp::LimitNode& limit) {
 }
 
 void ToGraph::addWrite(const lp::TableWriteNode& tableWrite) {
-  const auto writeKind =
-      static_cast<connector::WriteKind>(tableWrite.writeKind());
-  if (writeKind != connector::WriteKind::kInsert &&
-      writeKind != connector::WriteKind::kCreate) {
+  const auto writeKind = tableWrite.writeKind();
+  if (writeKind != WriteKind::kInsert && writeKind != WriteKind::kCreate) {
     VELOX_NYI("Only INSERT supported for TableWrite");
   }
   VELOX_CHECK_NULL(
